[Core, Order, Product, CoreBundle] refactor digital product: move all new fields to Core and CoreBundle
