+++ src/CoreShop/Bundle/CoreBundle/Resources/config/services/order.yml
@@ -13,6 +13,18 @@ services:
       - '@coreshop.core.order.transformer.cart_to_quote.inner'
       - '@coreshop.currency_converter'
 
+  coreshop.core.order.transformer.cart_item_to_order_item:
+    decorates: coreshop.order.transformer.cart_item_to_order_item
+    class: CoreShop\Component\Core\Order\Transformer\CartItemToSaleItemTransformer
+    arguments:
+      - '@coreshop.core.order.transformer.cart_item_to_order_item.inner'
+
+  coreshop.core.order.transformer.cart_item_to_quote_item:
+    decorates: coreshop.order.transformer.cart_item_to_quote_item
+    class: CoreShop\Component\Core\Order\Transformer\CartItemToSaleItemTransformer
+    arguments:
+      - '@coreshop.core.order.transformer.cart_item_to_quote_item.inner'
+
   coreshop.cart.cleanup.maintenance_listener:
     class: CoreShop\Bundle\CoreBundle\EventListener\CartCleanerMaintenanceManagerListener
     arguments:
