+++ src/CoreShop/Bundle/CoreBundle/Resources/install/pimcore/classes/CoreShopOrderBundle/CoreShopQuoteItem.json
@@ -1,6 +1,6 @@
 {
     "description": null,
-    "parentClass": "CoreShop\\Component\\Order\\Model\\QuoteItem",
+    "parentClass": "CoreShop\\Component\\Core\\Model\\QuoteItem",
     "useTraits": null,
     "allowInherit": false,
     "allowVariants": false,
