+++ src/CoreShop/Component/Core/Model/OrderItemInterface.php
@@ -8,22 +8,13 @@
  *
  * @copyright  Copyright (c) 2015-2017 Dominik Pfaffenbauer (https://www.pfaffenbauer.at)
  * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
- */
+*/
 
 namespace CoreShop\Component\Core\Model;
 
-use CoreShop\Component\Order\Model\CartItemInterface as BaseCartItemInterface;
-use CoreShop\Component\Shipping\Model\ShippableItemInterface;
+use CoreShop\Component\Order\Model\OrderItemInterface as BaseOrderItemInterface;
 
-interface CartItemInterface extends BaseCartItemInterface, ShippableItemInterface
+interface OrderItemInterface extends BaseOrderItemInterface, SaleItemInterface
 {
-    /**
-     * @return bool
-     */
-    public function getDigitalProduct();
 
-    /**
-     * @param bool $digitalProduct
-     */
-    public function setDigitalProduct($digitalProduct);
 }
