+++ src/CoreShop/Component/Core/Model/Product.php
@@ -58,6 +58,22 @@ class Product extends BaseProduct implements ProductInterface
     /**
      * {@inheritdoc}
      */
+    public function getDigitalProduct()
+    {
+        throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setDigitalProduct($digitalProduct)
+    {
+        throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function getEnabled()
     {
         return $this->getActive() && $this->getPublished();
