+++ src/CoreShop/Component/Order/Model/CartInterface.php
@@ -12,11 +12,9 @@
 
 namespace CoreShop\Component\Order\Model;
 
-
 use CoreShop\Component\Order\Checkout\CheckoutAwareInterface;
 use CoreShop\Component\Resource\Pimcore\Model\PimcoreModelInterface;
 use CoreShop\Component\StorageList\Model\StorageListInterface;
-
 
 interface CartInterface extends ProposalInterface, PimcoreModelInterface, CheckoutAwareInterface, StorageListInterface
 {
@@ -90,9 +88,4 @@ interface CartInterface extends ProposalInterface, PimcoreModelInterface, Checko
      * @return int
      */
     public function getPaymentFeeTaxRate();
-
-    /**
-     * @return bool
-     */
-    public function hasShippableItems();
 }
