+++ src/CoreShop/Component/Order/Model/CartItem.php
@@ -242,22 +242,6 @@ class CartItem extends AbstractPimcoreModel implements CartItemInterface
     /**
      * {@inheritdoc}
      */
-    public function getDigitalProduct()
-    {
-        throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setDigitalProduct($digitalProduct)
-    {
-        throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getTaxes()
     {
         throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
