+++ src/CoreShop/Component/Order/Model/CartItemInterface.php
@@ -28,16 +28,6 @@ interface CartItemInterface extends ProposalItemInterface, PimcoreModelInterface
     public function setIsGiftItem($isGiftItem);
 
     /**
-     * @return bool
-     */
-    public function getDigitalProduct();
-
-    /**
-     *  @param bool $digitalProduct
-     */
-    public function setDigitalProduct($digitalProduct);
-
-    /**
      * @return CartInterface
      */
     public function getCart();
