+++ src/CoreShop/Component/Product/Model/Product.php
@@ -103,22 +103,6 @@ class Product extends AbstractPimcoreModel implements ProductInterface
     /**
      * {@inheritdoc}
      */
-    public function getDigitalProduct()
-    {
-        throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setDigitalProduct($digitalProduct)
-    {
-        throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getName($language = null)
     {
         throw new ImplementedByPimcoreException(__CLASS__, __METHOD__);
